requirements.txt
torch>=2.0
transformers>=4.35
datasets
tokenizers
accelerate
bertopic
sentence-transformers
hdbscan
umap-learn
xgboost
scikit-learn
shap
imbalanced-learn
pandas
numpy
faker
streamlit
plotly
wordcloud
matplotlib
joblib
tqdm
pyarrow
